Added edit/delete authorization for restaurants/menu for individual users

+++ menu_app.py
@@ -1,42 +1,39 @@
+from database_setup import Base, Restaurant, MenuItem, User
 from flask import Flask, render_template, request, url_for,\
 	redirect, jsonify, flash, make_response
 from flask import session as login_session
-import random, string
-from sqlalchemy import create_engine
+from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Restaurant, MenuItem
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
-import httplib2
-import json
-import requests
+import random, string, json, requests, httplib2
 
-CLIENT_ID = json.loads(
-	open('client_secrets.json', 'r').read())['web']['client_id']
 
 app = Flask(__name__)
 
-# Connect to sql database and create session object
-engine = create_engine('sqlite:///restaurantmenu.db')
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Restaurant Menu Application"
+
+
+# Connect to Database and create database session
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
 
-@app.route('/')
-@app.route('/restaurants')
-def showRestaurants():
-	restaurants = session.query(Restaurant).all()
-	return render_template('restaurants.html', restaurants=restaurants)
-
-
+# Create anti-forgery state token
 @app.route('/login')
 def showLogin():
-	state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
-	login_session['state'] = state
+    state = ''.join(
+        random.choice(string.ascii_uppercase + string.digits) 
+        for x in range(32))
+    login_session['state'] = state
+    # return "The current session state is %s" % login_session['state']
+    return render_template('login.html', STATE=state)
 
-	return render_template('login.html', STATE=state)
 
 ## OAUTH CODE START ##
 # Revoke a current user's token and reset their login_session
@@ -44,11 +41,13 @@ def showLogin():
 def gconnect():
     # Validate state token
     if request.args.get('state') != login_session['state']:
-        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response = make_response(
+        	json.dumps('Invalid state parameter.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    # Obtain authorization code
-    code = request.data
+    # Obtain authorization code, now compatible with Python3
+    request.get_data()
+    code = request.data.decode('utf-8')
 
     try:
         # Upgrade the authorization code into a credentials object
@@ -65,11 +64,16 @@ def gconnect():
     access_token = credentials.access_token
     url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
            % access_token)
+    # Submit request, parse response - Python3 compatible
     h = httplib2.Http()
-    result = json.loads(h.request(url, 'GET')[1])
+    response = h.request(url, 'GET')[1]
+    str_response = response.decode('utf-8')
+    result = json.loads(str_response)
+
     # If there was an error in the access token info, abort.
     if result.get('error') is not None:
-        response = make_response(json.dumps(result.get('error')), 500)
+        response = make_response(
+        	json.dumps(result.get('error')), 500)
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -85,25 +89,24 @@ def gconnect():
     if result['issued_to'] != CLIENT_ID:
         response = make_response(
             json.dumps("Token's client ID does not match app's."), 401)
-        print "Token's client ID does not match app's."
         response.headers['Content-Type'] = 'application/json'
         return response
 
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(
+        	json.dumps('Current user is already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Store the access token in the session for later use.
-    login_session['access_token'] = credentials.access_token
+    login_session['access_token'] = access_token
     login_session['gplus_id'] = gplus_id
 
     # Get user info
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
-    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    params = {'access_token': access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
 
     data = answer.json()
@@ -112,6 +115,12 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+    # see if user exists, if it doesn't make a new one
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -119,38 +128,70 @@ def gconnect():
     output += '<img src="'
     output += login_session['picture']
     output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
-    flash("You are now logged in as %s" % login_session['username'])
-    print "done!"
+    flash("you are now logged in as %s" % login_session['username'])
     return output
+
+
+## LOCAL PERMISSIONS CODE START ##
+def createUser(login_session):
+    newUser = User(
+    	name=login_session['username'], 
+    	email=login_session['email'], 
+    	picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(
+    	email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(
+    	id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(
+        	email=email).one()
+        return user.id
+    except:
+        return None
+## LOCAL PERMISSION CODE END ##
 
 
 @app.route('/gdisconnect')
 def gdisconnect():
+        # Only disconnect a connected user.
     access_token = login_session.get('access_token')
     if access_token is None:
-        print 'Access Token is None'
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    print 'In gdisconnect access token is %s', access_token
-    print 'User name is: '
-    print login_session['username']
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' \
+    	% access_token
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
-    print 'result is '
-    print result
-    if result['status'] == '200':
+
+    if result.status == 200:
+        # Reset the user's sesson.
         del login_session['access_token']
         del login_session['gplus_id']
         del login_session['username']
         del login_session['email']
         del login_session['picture']
-        response = make_response(json.dumps('Successfully disconnected.'), 200)
+
+        response = make_response(
+        	json.dumps('Successfully disconnected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        # For whatever reason, the given token was invalid.
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
 ## OAUTH CODE END ##
@@ -169,15 +210,24 @@ def showMenuJSON(restaurant_id):
 	restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
 	items = session.query(MenuItem).filter_by(
 		restaurant_id=restaurant.id).all()
-
 	return jsonify(MenuItems = [i.serialize for i in items])
 
 
 @app.route('/restaurants/<int:restaurant_id>/menu/<int:menu_id>/JSON')
 def menuItemJSON(restaurant_id, menu_id):
 	menuItem = session.query(MenuItem).filter_by(id=menu_id).one()
-
 	return jsonify(MenuItem = menuItem.serialize)
+
+
+@app.route('/')
+@app.route('/restaurants')
+def showRestaurants():
+	restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
+	if 'username' not in login_session:
+		return render_template('publicrestaurants.html',
+								restaurants=restaurants)
+	else:
+		return render_template('restaurants.html', restaurants=restaurants)
 
 
 # Create new entry on Restaurant table, request form data from 
@@ -186,8 +236,10 @@ def menuItemJSON(restaurant_id, menu_id):
 def newRestaurant():
 	if 'username' not in login_session:
 		return redirect('/login')
+
 	if request.method == 'POST':
-		newRestaurant = Restaurant(name = request.form['name'])
+		newRestaurant = Restaurant(name = request.form['name'],
+								user_id = login_session['user_id'])
 		session.add(newRestaurant)
 		session.commit()
 		flash('New Restaurant Created')
@@ -208,6 +260,10 @@ def editRestaurant(restaurant_id):
 
 	if 'username' not in login_session:
 		return redirect('/login')
+
+	if editedRestaurant.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
+		
 	if request.method == 'POST':
 		if request.form['name']:
 			editedRestaurant.name = request.form['name']
@@ -231,6 +287,10 @@ def deleteRestaurant(restaurant_id):
 
 	if 'username' not in login_session:
 		return redirect('/login')
+
+	if deletedRestaurant.user_id != login_session['user_id']:
+		return "<script> function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
+
 	if request.method =='POST':
 		session.delete(deletedRestaurant)
 		session.commit()
@@ -248,9 +308,21 @@ def deleteRestaurant(restaurant_id):
 # for the restaurant id and another for the item's id
 @app.route('/restaurants/<int:restaurant_id>/menu')
 def showMenu(restaurant_id):
-	restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-	items = session.query(MenuItem).filter_by(restaurant_id=restaurant_id)
-	return render_template('menu.html', restaurant=restaurant, items=items)
+	restaurant = session.query(Restaurant).filter_by(
+		d=restaurant_id).one()
+	creator = getUserInfo(restaurant.user_id)
+	items = session.query(MenuItem).filter_by(
+		restaurant_id=restaurant_id).all()
+
+	if 'username' not in login_session or creator.id != login_session['user_id']:
+		return render_template('publicmenu.html', 
+			items=items, 
+			restaurant=restaurant, 
+			creator=creator)
+	else:
+		return render_template('menu.html', 
+			restaurant=restaurant, 
+			items=items)
 
 
 # Creates new MenuItem entry, pulls form data from newMenuItem.html
@@ -259,22 +331,28 @@ def showMenu(restaurant_id):
 def newMenuItem(restaurant_id):
 	if 'username' not in login_session:
 		return redirect('/login')
+	restaurant = session.query(Restaurant).filter_by(
+		id=restaurant_id).one()
 	if request.method == 'POST':
 		newItem = MenuItem(
 			name 			= request.form['name'],
 			description 	= request.form['desc'],
 			price 			= request.form['price'],
 			course 			= request.form['course'],
-			restaurant_id 	= restaurant_id)
+			restaurant_id 	= restaurant_id,
+			user_id 		= restaurant.user_id)
 
 		session.add(newItem)
 		session.commit()
 		flash('Menu Item Created')
 
-		return redirect(url_for('showMenu', restaurant_id=restaurant_id))
+		return redirect(url_for('showMenu', 
+						restaurant_id=restaurant_id)
+		)
 	else:
-		return render_template(
-			'newMenuItem.html', restaurant_id=restaurant_id)
+		return render_template('newMenuItem.html', 
+								restaurant_id=restaurant_id
+		)
 
 
 # Get the current values for a MenuItem entry, check all form data from 
@@ -282,26 +360,39 @@ def newMenuItem(restaurant_id):
 @app.route('/restaurants/<int:restaurant_id>/menu/<int:menu_id>/edit', 
 	methods=['GET', 'POST'])
 def editMenuItem(restaurant_id, menu_id):
-	editedItem = session.query(MenuItem).filter_by(
-		restaurant_id=restaurant_id, id=menu_id).one()
-
 	if 'username' not in login_session:
 		return redirect('/login')
+	editedItem = session.query(MenuItem).filter_by(
+		restaurant_id=restaurant_id, id=menu_id).one()
+	user = session.query(User).filter_by(
+		id = login_session['user_id']).one()
+	
+	if editedItem.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to edit this menu item. Please create your own menu item in order to edit.');}</script><body onload='myFunction()''>"
+
 	if request.method == 'POST':
-		if request.form['name']: editedItem.name = request.form['name']
-		if request.form['desc']: editedItem.description = request.form['desc']
-		if request.form['price']: editedItem.price 	= request.form['price']
-		if request.form['course']: editedItem.course = request.form['course']
+		if request.form['name']: 
+			editedItem.name = request.form['name']
+		if request.form['desc']: 
+			editedItem.description = request.form['desc']
+		if request.form['price']: 
+			editedItem.price 	= request.form['price']
+		if request.form['course']: 
+			editedItem.course = request.form['course']
 
 		session.add(editedItem)
 		session.commit()
 		flash('Menu Item Successfully Edited')
 
-		return redirect(url_for('showMenu', restaurant_id=restaurant_id))
+		return redirect(url_for('showMenu', 
+			restaurant_id=restaurant_id))
 
 	else:
-		return render_template(
-			'editMenuItem.html', restaurant_id=restaurant_id, item=editedItem)
+		return render_template('editmenuitem.html', 
+			restaurant_id=restaurant_id, 
+			menu_id=menu_id, 
+			item=editedItem)
+
 
 
 # Needs both restaurant id and menu item id to select corrent MenuItem entry
@@ -313,6 +404,10 @@ def deleteMenuItem(restaurant_id, menu_id):
 
 	if 'username' not in login_session:
 		return redirect('/login')
+	
+	if deletedItem.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to delete this menu item. Please create your own menu item in order to delete.');}</script><body onload='myFunction()''>"
+
 	if request.method == 'POST':
 		session.delete(deletedItem)
 		session.commit()
